pyproject.toml
[tool.poetry]
name = "api-watchbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "api_watchbot"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytelegrambotapi = "^4.12.0"
apscheduler = "^3.10.1"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
